NodePublisher, NodeSubscriber Tests: Rename to reflect class being tested
